# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This is a sample recipes API. You can find out more about the API at https://github.com/PacktPublishing/Building-Distributed-Applications-in-Gin.",
    "title": "Recipes API",
    "contact": {
      "name": "Mohamed Labouardy",
      "url": "https://labouardy.com",
      "email": "mohamed@labouardy.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/": {
      "get": {
        "description": "Index",
        "produces": [
          "application/json"
        ],
        "tags": [
          "index"
        ],
        "operationId": "indexPage"
      }
    },
    "/recipes": {
      "get": {
        "description": "Returns list of recipes",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "listRecipes"
      },
      "post": {
        "description": "Create a new recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "newRecipe",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Recipe"
            }
          }
        ]
      }
    },
    "/recipes/search": {
      "get": {
        "description": "Search recipes based on tags",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "findRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "recipe tag",
            "name": "tag",
            "in": "query",
            "required": true
          }
        ]
      }
    },
    "/recipes/{id}": {
      "get": {
        "description": "Get one recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "oneRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the recipe",
            "name": "id",
            "in": "path",
            "required": true
          }
        ]
      },
      "put": {
        "description": "Update an existing recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "updateRecipe",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Recipe"
            }
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ]
      },
      "delete": {
        "description": "Delete an existing recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "deleteRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the recipe",
            "name": "id",
            "in": "path",
            "required": true
          }
        ]
      }
    },
    "/refresh": {
      "post": {
        "description": "Refresh token",
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "operationId": "refresh"
      }
    },
    "/sign-in": {
      "post": {
        "description": "Login with username and password",
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "operationId": "signIn",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Password",
            "name": "password",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Username",
            "name": "username",
            "in": "query"
          }
        ]
      }
    },
    "/signout": {
      "post": {
        "description": "Signing out",
        "tags": [
          "auth"
        ],
        "operationId": "signOut"
      }
    }
  },
  "definitions": {
    "Recipe": {
      "type": "object",
      "properties": {
        "ingredients": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Ingredients"
        },
        "instructions": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Instructions"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Tags"
        }
      },
      "x-go-package": "github.com/mamalmaleki/go_recipe_api/models"
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}